Change NBT command to display new item component info
